moved source code from PatchingExperiment to Patching class
